Reuses ReferenceComponentConverter for FHIR Version compatibility
